Worked on width, height of profile-card.
Got the width and height of profile card right, made it responsive. Also added the measurements for header, middle and bottom parts of profile-card.

Next: Assign css shape for the profile picture, place the profile-pic in the middle of header and middle section. Also add padding or margin around it with the background-color of white. It has to be circular as well.

Next, make the p following h1 inline so that it gets placed in the same line as the h1.
Then, get the  P's to break inside profile-data